requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
elasticsearch==7.15.1
Flask==2.0.2
google-transliteration-api==1.0.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
numpy==1.21.2
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.2.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
